:boom: deprecate extraction module, refactor into existing image & pdf modules
