--- lib/mindee/extraction.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'extraction/tax_extractor'
-require_relative 'extraction/multi_receipts_extractor'
