--- lib/mindee/extraction/multi_receipts_extractor.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'multi_receipts_extractor/multi_receipts_extractor'
