--- lib/mindee/extraction/tax_extractor.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'tax_extractor/tax_extractor'
