+++ lib/mindee/image/image_extractor.rb
@@ -37,7 +37,7 @@ module Mindee
       # to extract.
       # @return [Array<Mindee::Image::ExtractedImage>] Extracted Images.
       def self.extract_multiple_images_from_source(input_source, page_id, polygons)
-        new_stream = load_input_source_pdf_page_as_image(input_source, page_id)
+        new_stream = load_input_source_pdf_page_as_stringio(input_source, page_id)
         new_stream.seek(0)
 
         extract_images_from_polygons(input_source, new_stream, page_id, polygons)
@@ -54,8 +54,8 @@ module Mindee
         extracted_elements = []
 
         polygons.each_with_index do |polygon, element_id|
-          polygon = normalize_polygon(polygon)
-          page_content = read_page_content(pdf_stream)
+          polygon = ImageUtils.normalize_polygon(polygon)
+          page_content = ImageUtils.read_page_content(pdf_stream)
 
           min_max_x = Geometry.get_min_max_x([
                                                polygon.top_left,
@@ -69,8 +69,8 @@ module Mindee
                                                polygon.top_right,
                                                polygon.bottom_left,
                                              ])
-          file_extension = determine_file_extension(input_source)
-          cropped_image = crop_image(page_content, min_max_x, min_max_y)
+          file_extension = ImageUtils.determine_file_extension(input_source)
+          cropped_image = ImageUtils.crop_image(page_content, min_max_x, min_max_y)
           if file_extension == 'pdf'
             cropped_image.format('jpg')
           else
@@ -78,70 +78,13 @@ module Mindee
           end
 
           buffer = StringIO.new
-          write_image_to_buffer(cropped_image, buffer)
+          ImageUtils.write_image_to_buffer(cropped_image, buffer)
           file_name = "#{input_source.filename}_page#{page_id}-#{element_id}.#{file_extension}"
 
           extracted_elements << create_extracted_image(buffer, file_name, page_id, element_id)
         end
 
         extracted_elements
-      end
-
-      # Retrieves the bounding box of a polygon.
-      #
-      # @param [Array<Point>, Mindee::Geometry::Polygon] polygon
-      def self.normalize_polygon(polygon)
-        if polygon.is_a?(Mindee::Geometry::Polygon)
-          Mindee::Geometry.get_bounding_box(polygon)
-        else
-          polygon
-        end
-      end
-
-      # Loads a buffer into a MiniMagick Image.
-      #
-      # @param [StringIO] pdf_stream Buffer containg the PDF
-      # @return [MiniMagick::Image] a valid MiniMagick image handle.
-      def self.read_page_content(pdf_stream)
-        pdf_stream.rewind
-        MiniMagick::Image.read(pdf_stream)
-      end
-
-      # Crops a MiniMagick Image from a the given bounding box.
-      #
-      # @param [MiniMagick::Image] image Input Image.
-      # @param [Mindee::Geometry::MinMax] min_max_x minimum & maximum values for the x coordinates.
-      # @param [Mindee::Geometry::MinMax] min_max_y minimum & maximum values for the y coordinates.
-      def self.crop_image(image, min_max_x, min_max_y)
-        width = image[:width].to_i
-        height = image[:height].to_i
-
-        image.format('jpg')
-        new_width = (min_max_x.max - min_max_x.min) * width
-        new_height = (min_max_y.max - min_max_y.min) * height
-        image.crop("#{new_width}x#{new_height}+#{min_max_x.min * width}+#{min_max_y.min * height}")
-
-        image
-      end
-
-      # Writes a MiniMagick::Image to a buffer.
-      #
-      # @param [MiniMagick::Image] image a valid MiniMagick image.
-      # @param [StringIO] buffer
-      def self.write_image_to_buffer(image, buffer)
-        image.write(buffer)
-      end
-
-      # Retrieves the file extension from the main file to apply it to the extracted images. Note: coerces pdf as jpg.
-      #
-      # @param [Mindee::Input::Source::LocalInputSource] input_source Local input source.
-      # @return [String] A valid file extension.
-      def self.determine_file_extension(input_source)
-        if input_source.pdf? || input_source.filename.downcase.end_with?('pdf')
-          'jpg'
-        else
-          File.extname(input_source.filename).strip.downcase[1..]
-        end
       end
 
       # Generates an ExtractedImage.
@@ -163,14 +106,38 @@ module Mindee
       #
       # @param input_file [LocalInputSource] Local input.
       # @param [Integer] page_id Page ID.
-      # @return [MiniMagick::Image] A valid PdfDocument handle.
-      def self.load_input_source_pdf_page_as_image(input_file, page_id)
+      # @return [StringIO] A valid PdfDocument handle.
+      def self.load_input_source_pdf_page_as_stringio(input_file, page_id)
         input_file.io_stream.rewind
         if input_file.pdf?
           Mindee::PDF::PdfProcessor.get_page(Origami::PDF.read(input_file.io_stream), page_id)
         else
           input_file.io_stream
         end
+      end
+
+      def self.extract_receipts(input_source, inference)
+        # Extracts individual receipts from multi-receipts documents.
+        #
+        # @param input_source [LocalInputSource] Local Input Source to extract sub-receipts from.
+        # @param inference [Inference] Results of the inference.
+        # @return [Array<ExtractedImage>] Individual extracted receipts as an array of ExtractedMultiReceiptsImage.
+
+        images = []
+        unless inference.prediction.receipts
+          raise Errors::MindeeInputError,
+                'No possible receipts candidates found for Multi-Receipts extraction.'
+        end
+
+        (0...input_source.count_pdf_pages).each do |page_id|
+          receipt_positions = inference.pages[page_id].prediction.receipts.map(&:bounding_box)
+          images.concat(
+            extract_multiple_images_from_source(input_source, page_id + 1,
+                                                receipt_positions)
+          )
+        end
+
+        images
       end
     end
   end
