+++ spec/extraction/multi_receipts_extractor_spec.rb
@@ -2,11 +2,9 @@
 
 require 'mindee/product'
 require 'mindee/input/sources'
-require 'mindee/extraction'
 require_relative '../data'
 
-describe Mindee::Extraction::MultiReceiptsExtractor do
-  include Mindee::Image
+describe 'multi-receipts extraction' do
   let(:empty_inference) do
     double('Inference', prediction: double('Prediction', receipts: nil), pages: [])
   end
@@ -39,7 +37,7 @@ describe Mindee::Extraction::MultiReceiptsExtractor do
       input_sample = Mindee::Input::Source::PathInputSource.new(multi_receipts_single_page_path)
       response = load_json(multi_receipts_single_page_json_path, 'complete.json')
       doc = Mindee::Product::MultiReceiptsDetector::MultiReceiptsDetectorV1.new(response['document']['inference'])
-      extracted_receipts = Mindee::Extraction::MultiReceiptsExtractor.extract_receipts(input_sample, doc)
+      extracted_receipts = Mindee::Image::ImageExtractor.extract_receipts(input_sample, doc)
 
       expect(extracted_receipts.size).to eq(6)
 
@@ -87,7 +85,7 @@ describe Mindee::Extraction::MultiReceiptsExtractor do
       input_sample = Mindee::Input::Source::PathInputSource.new(multi_receipts_multi_page_path)
       response = load_json(multi_receipts_multi_page_json_path, 'multipage_sample.json')
       doc = Mindee::Product::MultiReceiptsDetector::MultiReceiptsDetectorV1.new(response['document']['inference'])
-      extracted_receipts = Mindee::Extraction::MultiReceiptsExtractor.extract_receipts(input_sample, doc)
+      extracted_receipts = Mindee::Image::ImageExtractor.extract_receipts(input_sample, doc)
 
       expect(extracted_receipts.size).to eq(5)
 
@@ -126,7 +124,7 @@ describe Mindee::Extraction::MultiReceiptsExtractor do
   context 'when no receipts are found in inference' do
     it 'raises a MindeeInputError' do
       expect do
-        described_class.extract_receipts(empty_input_source, empty_inference)
+        Mindee::Image::ImageExtractor.extract_receipts(empty_input_source, empty_inference)
       end.to raise_error(Mindee::Errors::MindeeInputError,
                          'No possible receipts candidates found for Multi-Receipts extraction.')
     end
@@ -134,7 +132,8 @@ describe Mindee::Extraction::MultiReceiptsExtractor do
 
   context 'when input source has no pages' do
     it 'returns an empty array' do
-      extracted_receipts = described_class.extract_receipts(empty_input_source, valid_inference_with_no_receipts)
+      extracted_receipts = Mindee::Image::ImageExtractor.extract_receipts(empty_input_source,
+                                                                          valid_inference_with_no_receipts)
       expect(extracted_receipts).to eq([])
     end
   end
